Update verified-programs-status endpoint
